pyproject.toml
[tool.poetry]
name = "form-builder-backend"
version = "0.1.0"
description = ""
authors = ["Naman Malik <namanmalik06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2.0"
celery = "^5.3.6"
black = "^24.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
